fix(modal): move modal component to widgeta
