+++ src/app/widgets/modal/modal.component.less
