+++ src/app/widgets/modal/modal.component.ts
